# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates/prescriptions:
    get:
      operationId: listPrescriptionTemplates
      tags: [templates]
      summary: List prescription templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrescriptionTemplate"
    post:
      operationId: createPrescriptionTemplate
      tags: [templates]
      summary: Create a prescription template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionTemplate"
  /templates/prescriptions/{id}:
    patch:
      operationId: updatePrescriptionTemplate
      tags: [templates]
      summary: Update a prescription template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionTemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionTemplate"
    delete:
      operationId: deletePrescriptionTemplate
      tags: [templates]
      summary: Delete a prescription template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /templates/investigations:
    get:
      operationId: listInvestigationTemplates
      tags: [templates]
      summary: List investigation/lab test templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestigationTemplate"
    post:
      operationId: createInvestigationTemplate
      tags: [templates]
      summary: Create investigation template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestigationTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestigationTemplate"

  # ─────────────────────────────────────────────
  # USERS & ROLES
  # ─────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List system users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (freeze, change role, reset password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
  /roles:
    get:
      operationId: listRoles
      tags: [users]
      summary: List all roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [users]
      summary: Create a role with permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    patch:
      operationId: updateRole
      tags: [users]
      summary: Update role permissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"

  # ─────────────────────────────────────────────
  # SETTINGS
  # ─────────────────────────────────────────────
...
  /settings/referral-providers:
    get:
      operationId: listReferralProviders
      tags: [settings]
      summary: List referral providers
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralProvider"
    post:
      operationId: createReferralProvider
      tags: [settings]
      summary: Create a referral provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralProviderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralProvider"
  /settings/tax:
    get:
      operationId: getTaxSettings
      tags: [settings]
      summary: Get tax configuration
      responses:
        "200":
          description: Tax settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSettings"
    patch:
      operationId: updateTaxSettings
      tags: [settings]
      summary: Update tax configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSettings"
  /settings/system:
    get:
      operationId: getSystemSettings
      tags: [settings]
      summary: Get general system settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    patch:
      operationId: updateSystemSettings
      tags: [settings]
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    PrescriptionTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, content, createdAt]
    PrescriptionTemplateInput:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
      required: [name, content]
    PrescriptionTemplateUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
    InvestigationTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [labs, imaging, endoscopy, pathology]
        tests:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, type, tests, createdAt]
    InvestigationTemplateInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [labs, imaging, endoscopy, pathology]
        tests:
          type: array
          items:
            type: string
      required: [name, type, tests]

    # ── Users / Roles ──
    SystemUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        roleId:
          type: integer
        roleName:
          type: string
        branch:
          type: ["string", "null"]
        isFrozen:
          type: boolean
        createdAt:
          type: string
      required: [id, name, username, roleId, roleName, isFrozen, createdAt]
    SystemUserInput:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: ["string", "null"]
        roleId:
          type: integer
        branch:
          type: ["string", "null"]
      required: [name, username, password, roleId]
    SystemUserUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        roleId:
          type: ["integer", "null"]
        branch:
          type: ["string", "null"]
        isFrozen:
          type: ["boolean", "null"]
        password:
          type: ["string", "null"]
    Permissions:
      type: object
      properties:
        addPatient:
          type: boolean
        editPatient:
          type: boolean
        viewPatientFile:
          type: boolean
        listPatients:
          type: boolean
        deletePatient:
          type: boolean
        exportPatients:
          type: boolean
        addAppointment:
          type: boolean
        viewTodayAppointments:
          type: boolean
        viewAllAppointments:
          type: boolean
        editAppointment:
          type: boolean
        monitorAppointments:
          type: boolean
        addExpense:
          type: boolean
        viewExpenses:
          type: boolean
        deleteExpense:
          type: boolean
        viewFinancialReports:
          type: boolean
        payReceivables:
          type: boolean
        updateDiagnosis:
          type: boolean
        editMedicalRecord:
          type: boolean
        viewVisitHistory:
          type: boolean
        addVitals:
          type: boolean
        viewAnalytics:
          type: boolean
        manageInventory:
          type: boolean
        manageUsers:
          type: boolean
        manageSettings:
          type: boolean
      required: [addPatient, editPatient, viewPatientFile, listPatients, deletePatient, exportPatients, addAppointment, viewTodayAppointments, viewAllAppointments, editAppointment, monitorAppointments, addExpense, viewExpenses, deleteExpense, viewFinancialReports, payReceivables, updateDiagnosis, editMedicalRecord, viewVisitHistory, addVitals, viewAnalytics, manageInventory, manageUsers, manageSettings]
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
        createdAt:
          type: string
      required: [id, name, permissions, createdAt]
    RoleInput:
      type: object
      properties:
        name:
          type: string
        permissions:
          $ref: "#/components/schemas/Permissions"
      required: [name, permissions]
    RoleUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        permissions:
          $ref: "#/components/schemas/Permissions"

    # ── Settings ──
...
    ReferralProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [id, name]
    ReferralProviderInput:
      type: object
      properties:
        name:
          type: string
        specialty:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [name]
    TaxSettings:
      type: object
      properties:
        branch:
          type: string
        taxType:
          type: string
          enum: [on_request, automatic]
        taxTitle:
          type: string
        taxPercentage:
          type: number
      required: [branch, taxType, taxTitle, taxPercentage]
    TaxSettingsUpdate:
      type: object
      properties:
        taxType:
          type: ["string", "null"]
        taxTitle:
          type: ["string", "null"]
        taxPercentage:
          type: ["number", "null"]
    SystemSettings:
      type: object
      properties:
        activeBranch:
          type: string
        appointmentOrder:
          type: string
          enum: [by_time, by_creation]
        autoRefreshMinutes:
          type: integer
        displayBranch:
          type: ["string", "null"]
      required: [activeBranch, appointmentOrder, autoRefreshMinutes]
    SystemSettingsUpdate:
      type: object
      properties:
        activeBranch:
          type: ["string", "null"]
        appointmentOrder:
          type: ["string", "null"]
        autoRefreshMinutes:
          type: ["integer", "null"]
        displayBranch:
          type: ["string", "null"]