# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Permissions } from './permissions';

export interface Role {
  id: number;
  name: string;
  permissions: Permissions;
  createdAt: string;
}