# lib/api-zod/src/generated/types/systemSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SystemSettingsAppointmentOrder } from './systemSettingsAppointmentOrder';

export interface SystemSettings {
  activeBranch: string;
  appointmentOrder: SystemSettingsAppointmentOrder;
  autoRefreshMinutes: number;
  /** @nullable */
  displayBranch?: string | null;
}